Cross platform implementation.
Doesn't set memory limit, but kills process on exceeding memory limit (using goroutines).
