+++ cmd/test/test.go
@@ -5,12 +5,15 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/cp-tools/cpt/pkg/conf"
 	"github.com/cp-tools/cpt/utils"
 
 	"github.com/fatih/color"
+	"github.com/jedib0t/go-pretty/v6/table"
+	"github.com/jedib0t/go-pretty/v6/text"
 )
 
 // Test tests
@@ -150,4 +153,53 @@ func Test(submissionFilePath, checkerScript string,
 				timeLimit, memoryLimit)
 		}
 	}
+}
+
+func (verd *testExecDetails) prettyPrint(testIndex int) {
+	c := color.New(color.FgBlue, color.Bold).SprintFunc()
+
+	// Test: 4    Verdict: WA    Time: 32ms
+	fmt.Println(c("Test:"), testIndex, "\t"+c("Verdict:"), verd.verdict)
+	// Time: 32ms    Memory: 2000kb
+	verd.timeConsumed = verd.timeConsumed.Truncate(time.Millisecond)
+	memoryKB := fmt.Sprintf("%dkb", verd.memoryConsumed/1024)
+	fmt.Println(c("Time:"), verd.timeConsumed, "\t"+c("Memory:"), memoryKB)
+
+	if verd.failLog != "" {
+		// Fail: Could not execute checker
+		fmt.Println(c("Fail log:"), strings.TrimSpace(verd.failLog))
+	}
+
+	if verd.stderrLog != "" {
+		// Stderr:
+		// 1 2 3
+		// a b c
+		fmt.Println(c("Stderr:"), strings.TrimSpace(verd.stderrLog))
+	}
+
+	if verd.checkerLog != "" {
+		// Checker Log: Wrong answer, expected 3, found 4.
+		fmt.Println(c("Checker log:"), strings.TrimSpace(verd.checkerLog))
+	}
+
+	if strings.Contains(verd.verdict, "WA") {
+		t := table.NewWriter()
+		t.SetStyle(table.StyleLight)
+		t.Style().Options.DrawBorder = false
+
+		headerColor := text.Colors{text.FgBlue, text.Bold}
+		t.SetColumnConfigs([]table.ColumnConfig{
+			{Number: 1, AlignHeader: text.AlignCenter, ColorsHeader: headerColor, WidthMax: 40},
+			{Number: 2, AlignHeader: text.AlignCenter, ColorsHeader: headerColor, WidthMax: 40},
+			{Number: 3, AlignHeader: text.AlignCenter, ColorsHeader: headerColor, WidthMax: 40},
+		})
+
+		t.AppendHeader(table.Row{"INPUT", "EXPECTED", "OUTPUT"})
+		t.AppendRow(table.Row{verd.testDetails.truncInput,
+			verd.testDetails.truncExpected, verd.testDetails.truncOutput})
+
+		fmt.Println()
+		fmt.Println(t.Render())
+	}
+	fmt.Println()
 }
