+++ go.mod
@@ -4,8 +4,10 @@ go 1.14
 
 require (
 	github.com/AlecAivazis/survey/v2 v2.2.7
+	github.com/StackExchange/wmi v0.0.0-20190523213315-cbe66965904d // indirect
 	github.com/cp-tools/cpt-lib/v2 v2.1.4
 	github.com/fatih/color v1.10.0
+	github.com/go-ole/go-ole v1.2.5 // indirect
 	github.com/gosuri/uilive v0.0.4
 	github.com/inconshreveable/go-update v0.0.0-20160112193335-8152e7eb6ccf
 	github.com/jedib0t/go-pretty/v6 v6.0.6
@@ -13,10 +15,10 @@ require (
 	github.com/knadh/koanf v0.15.0
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/mitchellh/mapstructure v1.4.1
+	github.com/shirou/gopsutil v3.21.1+incompatible
 	github.com/spf13/cobra v1.1.1
 	github.com/stretchr/testify v1.7.0
 	golang.org/x/mod v0.4.1
-	golang.org/x/sys v0.0.0-20200828194041-157a740278f4
 	gopkg.in/yaml.v2 v2.4.0
 )
